cut useless dependencies
