+++ ydb/core/kqp/gateway/behaviour/tablestore/operations/ya.make
@@ -14,6 +14,7 @@ SRCS(
 PEERDIR(
     ydb/services/metadata/manager
     ydb/core/formats/arrow/serializer
+    ydb/core/tx/columnshard/engines/scheme/statistics/abstract
     ydb/core/kqp/gateway/utils
     ydb/core/protos
 )
