+++ ydb/core/tx/columnshard/blobs_action/abstract/ya.make
@@ -17,7 +17,7 @@ PEERDIR(
     ydb/core/protos
     contrib/libs/apache/arrow
     ydb/core/tablet_flat
-    ydb/core/tx/tiering
+    ydb/core/tx/tiering/abstract
     ydb/core/tx/columnshard/data_sharing/protos
     ydb/core/tx/columnshard/blobs_action/events
     ydb/core/tx/columnshard/blobs_action/protos
