+++ ydb/core/tx/columnshard/blobs_action/tier/storage.h
@@ -32,7 +32,7 @@ protected:
     virtual std::shared_ptr<IBlobsReadingAction> DoStartReadingAction() override;
     virtual std::shared_ptr<IBlobsGCAction> DoStartGCAction(const std::shared_ptr<TRemoveGCCounters>& counters) const override;
     virtual bool DoLoad(IBlobManagerDb& dbBlobs) override;
-    virtual void DoOnTieringModified(const std::shared_ptr<NColumnShard::TTiersManager>& tiers) override;
+    virtual void DoOnTieringModified(const std::shared_ptr<NColumnShard::ITiersManager>& tiers) override;
 
 public:
     TOperator(const TString& storageId, const NColumnShard::TColumnShard& shard, const std::shared_ptr<NDataSharing::TStorageSharedBlobsManager>& storageSharedBlobsManager);
