+++ ydb/core/tx/columnshard/engines/portions/with_blobs.h
@@ -108,7 +108,7 @@ public:
         return PortionInfo.BlobsBytes();
     }
 
-    void FillStatistics(const std::map<NStatistics::TIdentifier, NStatistics::TOperatorContainer>& operators, const TIndexInfo& index);
+    void FillStatistics(const TIndexInfo& index);
 
     static TPortionInfoWithBlobs BuildByBlobs(std::vector<TSplittedBlob>&& chunks,
         std::shared_ptr<arrow::RecordBatch> batch, const ui64 granule, const TSnapshot& snapshot, const std::shared_ptr<IStoragesManager>& operators,
