+++ ydb/core/tx/columnshard/engines/scheme/index_info.cpp
@@ -1,7 +1,6 @@
 #include "index_info.h"
 #include "statistics/abstract/operator.h"
 
-#include <ydb/core/tx/columnshard/engines/portions/with_blobs.h>
 #include <ydb/core/formats/arrow/arrow_batch_builder.h>
 #include <ydb/core/formats/arrow/sort_cursor.h>
 #include <ydb/core/sys_view/common/schema.h>
@@ -10,9 +9,6 @@
 #include <ydb/core/base/appdata.h>
 
 namespace NKikimr::NOlap {
-
-const TString TIndexInfo::STORE_INDEX_STATS_TABLE = TString("/") + NSysView::SysPathName + "/" + NSysView::StorePrimaryIndexStatsName;
-const TString TIndexInfo::TABLE_INDEX_STATS_TABLE = TString("/") + NSysView::SysPathName + "/" + NSysView::TablePrimaryIndexStatsName;
 
 static std::vector<TString> NamesOnly(const std::vector<TNameTypeInfo>& columns) {
     std::vector<TString> out;
@@ -317,12 +313,6 @@ TColumnSaver TIndexInfo::GetColumnSaver(const ui32 columnId, const TSaverContext
     }
 }
 
-std::shared_ptr<TColumnLoader> TIndexInfo::GetColumnLoaderVerified(const ui32 columnId) const {
-    auto result = GetColumnLoaderOptional(columnId);
-    AFL_VERIFY(result);
-    return result;
-}
-
 std::shared_ptr<TColumnLoader> TIndexInfo::GetColumnLoaderOptional(const ui32 columnId) const {
     auto it = ColumnFeatures.find(columnId);
     if (it == ColumnFeatures.end()) {
@@ -487,10 +477,6 @@ void TIndexInfo::InitializeCaches(const std::shared_ptr<IStoragesManager>& opera
         AFL_VERIFY(ArrowColumnByColumnIdCache.emplace(cId, GetColumnFieldVerified(cId)).second);
         AFL_VERIFY(ColumnFeatures.emplace(cId, TColumnFeatures::BuildFromIndexInfo(cId, *this, operators->GetDefaultOperator())).second);
     }
-}
-
-void TIndexInfo::FillStatistics(TPortionInfoWithBlobs& portion) const {
-    portion.FillStatistics(Statistics, *this);
 }
 
 } // namespace NKikimr::NOlap
