+++ ydb/core/tx/columnshard/engines/scheme/statistics/max/ya.make
@@ -7,7 +7,8 @@ SRCS(
 
 PEERDIR(
     ydb/core/tx/columnshard/engines/scheme/statistics/abstract
-    ydb/core/tx/schemeshard/olap/schema
+    ydb/core/tx/columnshard/engines/scheme/abstract
+    ydb/core/tx/columnshard/splitter/abstract
     ydb/core/formats/arrow
 )
