+++ ydb/core/tx/columnshard/engines/scheme/ya.make
@@ -16,6 +16,7 @@ PEERDIR(
     ydb/library/actors/core
     ydb/core/tx/columnshard/engines/scheme/indexes
     ydb/core/tx/columnshard/engines/scheme/statistics
+    ydb/core/tx/columnshard/engines/scheme/abstract
     ydb/core/tx/columnshard/blobs_action/abstract
 )
