+++ ydb/core/tx/columnshard/splitter/abstract/chunks.cpp
@@ -1,0 +1,5 @@
+#include "chunks.h"
+
+namespace NKikimr::NOlap {
+
+}
