+++ ydb/core/tx/columnshard/splitter/abstract/ya.make
@@ -1,0 +1,17 @@
+LIBRARY()
+
+SRCS(
+    chunks.cpp
+    chunk_meta.cpp
+)
+
+PEERDIR(
+    ydb/core/tx/columnshard/engines/scheme/abstract
+    ydb/core/tx/columnshard/common
+)
+
+END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
