+++ ydb/core/tx/tiering/abstract/ya.make
@@ -1,0 +1,11 @@
+LIBRARY()
+
+SRCS(
+    manager.cpp
+)
+
+PEERDIR(
+    ydb/library/actors/core
+)
+
+END()
