+++ ydb/services/metadata/common/ya.make
@@ -9,7 +9,7 @@ PEERDIR(
     ydb/services/metadata/initializer
     ydb/services/metadata/abstract
     ydb/services/bg_tasks/abstract
-    ydb/core/tx/schemeshard
+    ydb/core/tx/scheme_cache
 )
 
 YQL_LAST_ABI_VERSION()
